Rename local strings.[ch] to local_strings.[ch]
This avoids linker and human confusion with
the <usr/strings.h> include file. No functional
change.
