+++ usr/local_strings.h
